+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CRUDs.
 - BlueNaas CRUD tools
 
+### Fixed
+- Migrate LLM Evaluation logic to scripts and add tests
+
 ## [0.3.3] - 30.10.2024
 
 ### Changed